backend: initial work to add more backends
