+++ pywal/backends/__init__.py
@@ -1,0 +1,16 @@
+"""
+                                      '||
+... ...  .... ... ... ... ...  ....    ||
+ ||'  ||  '|.  |   ||  ||  |  '' .||   ||
+ ||    |   '|.|     ||| |||   .|' ||   ||
+ ||...'     '|       |   |    '|..'|' .||.
+ ||      .. |
+''''      ''
+Created by Dylan Araps.
+"""
+
+from . import wal
+
+__all__ = [
+    "wal",
+]
